+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,40 @@
 # demo.material.resrc.ch
 
+## 0.0.19
+
+### Patch Changes
+
+- Updated dependencies
+  - material.resrc.ch@0.0.19
+
+## 0.0.18
+
+### Patch Changes
+
+- Updated dependencies
+  - material.resrc.ch@0.0.18
+
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies
+  - material.resrc.ch@0.0.17
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - material.resrc.ch@0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies
+  - material.resrc.ch@0.0.15
+
 ## 0.0.14
 
 ### Patch Changes